+++ src/features/marketing/components/banner/Banner.component.tsx
