+++ src/features/marketing/components/banner/Banner.data.ts
