+++ src/features/marketing/components/banner/Banner.styles.scss
