+++ src/features/marketing/components/full-width-banner/FullWidthBanner.styles.scss
