+++ src/features/marketing/components/scroll-item/ScrollItem.styles.scss
