+++ src/features/marketing/components/scroll-list/ScrollList.helpers.ts
