+++ src/features/marketing/components/split-banner/SplitBanner.styles.scss
