+++ src/features/marketing/index.ts
@@ -1,0 +1,6 @@
+export { default as Banner } from "./components/banner/Banner.component";
+export { default as FullWidthBanner } from "./components/full-width-banner/FullWidthBanner.component";
+export { default as SplitBanner } from "./components/split-banner/SplitBanner.component";
+export { default as ScrollList } from "./components/scroll-list/ScrollList.component";
+
+export * from "./types/marketing.types";
