+++ src/features/marketing/types/marketing.types.ts
@@ -1,0 +1,4 @@
+export enum SCROLL_TYPE {
+  newArrival = "newArrival",
+  onSale = "onSale",
+}
